AffichageOk

+++ src/Erratique.java
@@ -11,8 +11,8 @@ public class Erratique extends Neuneu implements ObjetDessinable{
 	}
 	
 	void bougeDeLa(){
-		int deplacement = (int)Math.random()*9;
-		this.setMaCase(maison.getCase( getLaCase().getWPosition() + deplacement/3, getLaCase().getHPosition() + deplacement%3));
+		int deplacement = (int) Math.random()*9;
+		this.setMaCase(maison.getCase( getLaCase().getWPosition() + (int) (deplacement/3), getLaCase().getHPosition() + (int) (deplacement%3)));
 	}
 	void tour(){
 		bougeDeLa();
